requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
av>=10.0.0
numpy>=1.24.0
streamlit>=1.28.0
plotly>=5.15.0
Pillow>=9.5.0
